==> code/03. Extract data.ipynb <==
import os
import copy
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

dataset = pd.read_csv("../../data/BG dataset.csv")

dataset

patients = dataset.file_name.unique()
np.random.shuffle(patients)
patients_fold = np.array_split(patients, 10)

train_fold = []
test_fold = []
for i in range(10):
    fold = copy.deepcopy(patients_fold)
    test_fold.append(fold[i].tolist())
    del fold[i]
    train_fold.append(np.concatenate(fold).tolist())

np.savez("../../data/BG_Fold_Info.npz", train=train_fold, test=test_fold)

from sklearn.model_selection import TimeSeriesSplit
from scipy.stats import skew, kurtosis

#import tsfresh
#from tsfresh.feature_extraction import ComprehensiveFCParameters
#from tsfresh.utilities.dataframe_functions import impute

def gen_data(glucose_data, train_size, gap):
    
    tscv = TimeSeriesSplit(n_splits = len(glucose_data)-train_size-gap,
                           max_train_size = train_size,
                           test_size = 1,
                           gap = gap)
    train_list=[]
    test_list=[]
    
    for train_index, test_index in tscv.split(glucose_data):
        train, test = glucose_data[train_index], glucose_data[test_index]
        train_list.append(train.tolist())
        test_list.append(test.tolist()[0])
    
    return(train_list, test_list)

directory = "../../data/BG Ex data/"

if not os.path.exists(directory):
    os.makedirs(directory)

for fn in dataset.file_name.unique():
    glucose_data = dataset[dataset["file_name"]==fn].reset_index(drop=True)
    
    gen12_x, gen12_y = gen_data(glucose_data.glucose, 12, 5)
    gen24_x, gen24_y = gen_data(glucose_data.glucose, 24, 5)
    gen36_x, gen36_y = gen_data(glucose_data.glucose, 36, 5)
    
    gen12 = pd.DataFrame()
    for i, ts in enumerate(gen12_x):
        data = [[x,i] for x in ts]
        gen12 = gen12.append(data, ignore_index=True)
    gen12.columns=["value", "id"]
    
    gen24 = pd.DataFrame()
    for i, ts in enumerate(gen24_x):
        data = [[x,i] for x in ts]
        gen24 = gen24.append(data, ignore_index=True)
    gen24.columns=["value", "id"]
    
    gen36 = pd.DataFrame()
    for i, ts in enumerate(gen36_x):
        data = [[x,i] for x in ts]
        gen36 = gen36.append(data, ignore_index=True)
    gen36.columns=["value", "id"]
    
    #extraction_settings = ComprehensiveFCParameters()
    
    #tf=tsfresh.extract_features(gen12, 
    #                            column_id="id",
    #                            default_fc_parameters=extraction_settings,
    #                            impute_function=impute)
    #tf12 = tf.T.drop_duplicates().T
    
    #tf=tsfresh.extract_features(gen24, 
    #                            column_id="id",
    #                            default_fc_parameters=extraction_settings,
    #                            impute_function=impute)
    #tf24 = tf.T.drop_duplicates().T
    
    #tf=tsfresh.extract_features(gen36, 
    #                            column_id="id",
    #                            default_fc_parameters=extraction_settings,
    #                            impute_function=impute)
    #tf36 = tf.T.drop_duplicates().T
    
    np.savez_compressed("../../data/BG Ex data/"+fn+".npz",
                        gen12 = gen12_x[24:],
                        #tf12 = tf12[24:],
                        y12 = gen12_y[24:],
                        #tf12_name = tf12.columns.tolist(),
                        gen24 = gen24_x[12:],
                        #tf24 = tf24[12:],
                        y24 = gen24_y[12:],
                        #tf24_name = tf24.columns.tolist(),
                        gen36 = gen36_x,
                        #tf36 = tf36,
                        y36 = gen36_y#,
                        #tf36_name = tf36.columns.tolist()
                       )
