==> code/06. LR Based Stacking.ipynb <==
import os
import gc
import joblib
import warnings
warnings.filterwarnings('ignore')

import pandas as pd
import numpy as np
import statsmodels.api as sm

from sklearn.feature_selection import SequentialFeatureSelector as SFS
from sklearn.linear_model import LinearRegression
from sklearn.linear_model import Lasso

class Pred_stacking:
    
    def __init__(self, FOLD):
        
        self.train_x = None
        self.train_y = None
        
        self.test_x = None
        self.test_y = None
        
        self.train_pred_list = []
        self.test_pred_list = []
        
        self.FOLD = FOLD
        
    def __del__(self):
        
        print("Remove Success\n")
        
    def Add_data(self):
        
        train_data = pd.read_csv("../../data/BG Pred/Train_Fold"+str(self.FOLD)+"_pred.csv")
        test_data = pd.read_csv("../../data/BG Pred/Test_Fold"+str(self.FOLD)+"_pred.csv")
        
        self.train_x = train_data.drop("label", axis=1)
        self.train_y = train_data[["label"]]

        self.test_x = test_data.drop("label", axis=1)
        self.test_y = test_data[["label"]]
        
    def DF_concat_save(self):
        
        self.train_pred_list.append(self.train_y)
        self.test_pred_list.append(self.test_y)
        
        train_pred_df = pd.concat(self.train_pred_list, axis=1) 
        test_pred_df = pd.concat(self.test_pred_list, axis=1)
        
        train_pred_df.to_csv("../../data/BG LBStack/Train_Fold"+str(self.FOLD)+"_lbstack.csv", index=False)
        test_pred_df.to_csv("../../data/BG LBStack/Test_Fold"+str(self.FOLD)+"_lbstack.csv", index=False)
        
        print("Fold"+str(self.FOLD)+" file save !")
        
    def LR(self):
        
        lr = LinearRegression()
        lr.fit(self.train_x, self.train_y.values.flatten())
        
        lr_train_pred = lr.predict(self.train_x)
        lr_test_pred = lr.predict(self.test_x)
        
        self.train_pred_list.append(pd.DataFrame(lr_train_pred, columns=["LR"]))
        self.test_pred_list.append(pd.DataFrame(lr_test_pred, columns=["LR"]))
        
        joblib.dump(lr, "../../LBStack model/Fold"+str(self.FOLD)+"/LR.pkl")
        print("LR learning Success !")
        
    def LR_forward(self):
        
        lr = LinearRegression()
        sfs = SFS(lr, direction="forward")
        sfs.fit(self.train_x, self.train_y.values.flatten())
        
        select_col = list(sfs.get_feature_names_out())
        
        fw = LinearRegression()
        fw.fit(
            self.train_x[select_col],
            self.train_y.values.flatten()
        )
        
        fw_train_pred = fw.predict(self.train_x[select_col])
        fw_test_pred = fw.predict(self.test_x[select_col])
        
        self.train_pred_list.append(
            pd.DataFrame(fw_train_pred, columns=["LR forward"])
        )
        self.test_pred_list.append(
            pd.DataFrame(fw_test_pred, columns=["LR forward"])
        )
        
        joblib.dump(
            fw,
            "../../LBStack model/Fold"+str(self.FOLD)+"/LR forward.pkl"
        )
        joblib.dump(
            select_col,
            "../../LBStack model/Fold"+str(self.FOLD)+"/LR forward.txt"
        )
        print("LR forward learning Success !")        
        
    def LR_backward(self):
        
        lr = LinearRegression()
        sfs = SFS(lr, direction="backward")
        sfs.fit(self.train_x, self.train_y.values.flatten())
        
        select_col = list(sfs.get_feature_names_out())

        bw = LinearRegression()
        bw.fit(
            self.train_x[select_col],
            self.train_y.values.flatten()
        )
        
        bw_train_pred = bw.predict(self.train_x[select_col])
        bw_test_pred = bw.predict(self.test_x[select_col])
        
        self.train_pred_list.append(
            pd.DataFrame(bw_train_pred, columns=["LR backward"])
        )
        self.test_pred_list.append(
            pd.DataFrame(bw_test_pred, columns=["LR backward"])
        )
        
        joblib.dump(
            bw,
            "../../LBStack model/Fold"+str(self.FOLD)+"/LR backward.pkl"
        )
        joblib.dump(
            select_col,
            "../../LBStack model/Fold"+str(self.FOLD)+"/LR backward.txt"
        )
        print("LR backward learning Success !")        
        
    def LASSO(self):
        
        lasso = Lasso()
        lasso.fit(self.train_x, self.train_y.values.flatten())
        
        lasso_train_pred = lasso.predict(self.train_x)
        lasso_test_pred = lasso.predict(self.test_x)
        
        self.train_pred_list.append(
            pd.DataFrame(lasso_train_pred, columns=["LASSO"])
        )
        self.test_pred_list.append(
            pd.DataFrame(lasso_test_pred, columns=["LASSO"])
        )
        
        joblib.dump(lasso, "../../LBStack model/Fold"+str(self.FOLD)+"/LASSO.pkl")
        print("LASSO learning Success !")

directory = "../../data/BG LBStack/"
if not os.path.exists(directory):
    os.makedirs(directory)

directory = "../../LBStack model/"
if not os.path.exists(directory):
    os.makedirs(directory)
    
directory = "../../LBStack model/Fold"
for i in range(10):
    if not os.path.exists(directory+str(i)+"/"):
        os.makedirs(directory+str(i)+"/")

for k in range(10):
    
    PS = Pred_stacking(FOLD=k)
    
    PS.Add_data()
    
    PS.LR()
    PS.LASSO()
            
    PS.DF_concat_save()
    
    del(PS)
    gc.collect()
