==> code/08. DL90 Average.ipynb <==
import pandas as pd
import numpy as np

from sklearn.metrics import mean_squared_error

train_data, test_data = [], []

for i in range(10):
    train = pd.read_csv("../../data/BG Pred/Train_Fold"+str(i)+"_pred.csv")
    test = pd.read_csv("../../data/BG Pred/Test_Fold"+str(i)+"_pred.csv")
    
    train_data.append(train)
    test_data.append(test)

rm_col = ["label"]

test_x = test_data[0].drop(rm_col, axis=1)
test_y = test_data[0][["label"]]

import os
path = "../../data/BG Avg/"
if not os.path.exists(path):
    os.mkdir(path)

for i in range(10):
    train_x = train_data[i].drop(rm_col, axis=1)
    train_y = train_data[i][["label"]]
    
    test_x = test_data[i].drop(rm_col, axis=1)
    test_y = test_data[i][["label"]]
    
    train_pred = (train_x.sum(axis=1)/(91-len(rm_col))).tolist()
    test_pred = (test_x.sum(axis=1)/(91-len(rm_col))).tolist()
    
    train_pred_list, test_pred_list = [],[]
    
    train_pred_list.append(pd.DataFrame(train_pred, columns=["AVG"]))
    test_pred_list.append(pd.DataFrame(test_pred, columns=["AVG"]))
    
    train_pred_list.append(train_y)
    test_pred_list.append(test_y)
        
    train_pred_df = pd.concat(train_pred_list, axis=1) 
    test_pred_df = pd.concat(test_pred_list, axis=1)
        
    train_pred_df.to_csv("../../data/BG Avg/Train_Fold"+str(i)+"_avg.csv", index=False)
    test_pred_df.to_csv("../../data/BG Avg/Test_Fold"+str(i)+"_avg.csv", index=False)
